refactor: extract useGreetedMessages to keep AgentTabPanel under line budget
The main merge grew AgentTabPanel past the 500-line eslint budget; move the
greeting/intro hook into its own module.
