+++ web_src/src/components/CanvasToolSidebar/AgentTabPanel.tsx
@@ -22,7 +22,6 @@ import {
   AGENT_BOOT_CONTEXT_READY_EVENT,
   clearAgentBootContext,
   clearAgentBootContextForCanvas,
-  getAgentBootInitialMessage,
   getAgentBootMessage,
   isAgentBootReady,
 } from "@/lib/agentBootContext";
@@ -38,6 +37,7 @@ import {
 import type { AgentMessage, AgentOutgoingImage } from "./types";
 import type { CanvasToolSidebarState } from "./useCanvasToolSidebarState";
 import { groupMessages } from "./agentMessageGroups";
+import { useGreetedMessages } from "./useGreetedMessages";
 import { AgentSetupNotice } from "./AgentSetupState";
 import { getAgentSetupState } from "./agentSetupStateModel";
 
@@ -221,42 +221,6 @@ function ChatConversation({
       />
     </div>
   );
-}
-
-// Prepend a synthetic greeting (and optional template intro) so it never disappears.
-function useGreetedMessages(rawMessages: AgentMessage[], canvasId: string): AgentMessage[] {
-  const { account } = useContext(AccountContext);
-  const greetingFirstName = account?.name?.split(" ")[0] ?? "there";
-  // Read fresh each render so a /clear drops the intro once cleared for this canvas.
-  const bootInitialMessage = getAgentBootInitialMessage(canvasId);
-
-  return useMemo(() => {
-    const greeting: AgentMessage = {
-      id: "__greeting__",
-      role: "assistant",
-      content: `Hi ${greetingFirstName}! I'm your SuperPlane agent. I'll help you build and modify this canvas.`,
-      createdAt: rawMessages[0]?.createdAt ?? null,
-      toolCallId: "",
-      toolName: "",
-      toolStatus: "",
-    };
-
-    if (!bootInitialMessage) {
-      return [greeting, ...rawMessages];
-    }
-
-    const templateIntro: AgentMessage = {
-      id: "__boot_initial_message__",
-      role: "assistant",
-      content: bootInitialMessage,
-      createdAt: rawMessages[0]?.createdAt ?? null,
-      toolCallId: "",
-      toolName: "",
-      toolStatus: "",
-    };
-
-    return [greeting, templateIntro, ...rawMessages];
-  }, [rawMessages, greetingFirstName, bootInitialMessage]);
 }
 
 function resolveComposerStatusLabel(resetPending: boolean, sendPending: boolean, status: string): string {
